% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseDeg.R
\name{analyseDeg}
\alias{analyseDeg}
\title{Analyse degree or centrality}
\usage{
analyseDeg(networks, toolnames, n, mode = "hub")
}
\arguments{
\item{networks}{list of list of networks per tool}

\item{toolnames}{list of toolnames}

\item{n}{number of replicates}

\item{mode}{"hub" or "central": for hub, the degree is returned, for central, the centrality scores}
}
\value{
dataframe specifying tool ID and degree
}
\description{
Analyses degree or centrality for a list of outputs from different tools
}
\details{
Outputs a dataframe with degrees for all tools.
}
