% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseKlemm.R
\name{analyseKlemm}
\alias{analyseKlemm}
\title{Analyse Klemm matrices}
\usage{
analyseKlemm(klemmadj, n)
}
\arguments{
\item{klemmadj}{list of Klemm matrices}

\item{n}{number of replicates}

\item{setname}{name of simulation}
}
\value{
Dataframe with statistics for Klemm-Eguiluz matrices
}
\description{
Outputs network statistics for a list of Klemm-Eguiluz matrices.
}
\details{
Outputs assortativity, degree, transitivity and scale-freeness for Klemm-Eguiluz true positive matrices.
}
