% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callTools.R
\name{callTools}
\alias{callTools}
\title{Call tools to infer networks}
\usage{
callTools(alldata, klemmadj, toolnames, setname, x, n, absolute = FALSE,
  mode = "env")
}
\arguments{
\item{alldata}{list of datasets with simulated counts}

\item{klemmadj}{list of Klemm matrices}

\item{toolnames}{list of toolnames}

\item{setname}{name of simulation}

\item{x}{vector specificying environmental strength or removed species}

\item{n}{number of replicates}

\item{absolute}{if true: only check for edge presence, not matching sign}

\item{mode}{"env" or "abundance", env takes environmental strength into account while "abundance" includes species removal}
}
\value{
Returns nothing, but saves .rds files of networks and statistics
}
\description{
Infer networks using SpiecEasi, Spearman or gCoda
}
\details{
Calls on original implementations of SpiecEasi and gCoda or the CoNet implementation of Spearman. The wrapper also ensures that adjacency matrices are returned that are compatible with other functions.
}
