% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAdjacency.R
\name{computeAdjacency}
\alias{computeAdjacency}
\title{Compute adjacency}
\usage{
computeAdjacency(matrix)
}
\arguments{
\item{matrix}{matrix with values, e.g. Spearman correlation or other tool output}
}
\value{
Adjacency matrix
}
\description{
Computes adjacency matrix with only 1, -1 and 0 values
}
\details{
Compute adjacency matrix from a correlation or covariance matrix with only significant edges
}
