% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAUC.R
\name{computeAUC}
\alias{computeAUC}
\title{Compute AUC}
\usage{
computeAUC(imatrix, outmatrix)
}
\arguments{
\item{matrix}{matrix with values, e.g. Spearman correlation or other tool output}
}
\value{
AUC score
}
\description{
Calculates AUC coefficient from matrix
}
\details{
Wrapper for pROC function that converts matrix to igraph object
}
