% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCentral.R
\name{computeCentral}
\alias{computeCentral}
\title{Compute centrality}
\usage{
computeCentral(matrix)
}
\arguments{
\item{matrix}{matrix with values, e.g. Spearman correlation or other tool output}
}
\value{
top 5 betweenness centrality species
}
\description{
Calculates betweenness centrality from matrix
}
\details{
Returns the top 5 highest betweenness centrality nodes
}
