% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHub.R
\name{computeHub}
\alias{computeHub}
\title{Compute hubs}
\usage{
computeHub(matrix)
}
\arguments{
\item{matrix}{matrix with values, e.g. Spearman correlation or other tool output}
}
\value{
top 5 hub species
}
\description{
Identifies top 5 hub species from matrix
}
\details{
Returns the top 5 highest degree species
}
