% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePrecision.R
\name{computePrecision}
\alias{computePrecision}
\title{Compute precision}
\usage{
computePrecision(imatrix, outmatrix, absolute = FALSE)
}
\arguments{
\item{imatrix}{true positive matrix, e.g. Klemm-Eguiluz adjacency matrix}

\item{outmatrix}{matrix with values, e.g. Spearman correlation or other tool output}

\item{absolute}{Calculates matches for absolute values if true instead of taking sign into account}
}
\value{
precision score
}
\description{
Calculates precision of tools
}
\details{
Compares tool adjacency matrix output to Klemm-Eguiluz adjacency matrices
}
