% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpecificity.R
\name{computeSpecificity}
\alias{computeSpecificity}
\title{Compute specificity}
\usage{
computeSpecificity(imatrix, outmatrix, absolute = FALSE)
}
\arguments{
\item{imatrix}{true positive matrix, e.g. Klemm-Eguiluz adjacency matrix}

\item{outmatrix}{matrix with values, e.g. Spearman correlation or other tool output}

\item{absolute}{Calculates matches for absolute values if true instead of taking sign into account}
}
\value{
specificity score
}
\description{
Calculates specificity of tools
}
\details{
Compares tool adjacency matrix output to Klemm-Eguiluz adjacency matrices
}
