% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStat.R
\name{computeStat}
\alias{computeStat}
\title{Compute assortativity, transitivity or degree}
\usage{
computeStat(matrix, mode)
}
\arguments{
\item{matrix}{matrix with values, e.g. Spearman correlation or other tool output}
}
\value{
assortativity score
}
\description{
Calculates assortativity coefficient from matrix
}
\details{
Wrapper for igraph function that converts matrix to igraph object
}
