% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fileUploader.R
\name{fileUploader}
\alias{fileUploader}
\alias{mod_fileUploader_ui}
\alias{mod_fileUploader_server}
\title{fileUploader UI Function}
\usage{
mod_fileUploader_ui(id, label = "Upload a File", ...)

mod_fileUploader_server(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{label}{message show before upload a file}

\item{...}{any other argument for `shiny::fileInput`}
}
\value{
bs4Dash UI

reactive igraph object
}
\description{
A shiny Module.
}
