% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKlemms.R
\name{generateKlemms}
\alias{generateKlemms}
\title{Generate Klemm adjacency matrices}
\usage{
generateKlemms(n, s, pep, c)
}
\arguments{
\item{n}{number of replicates}

\item{s}{number of species}

\item{pep}{positive edge percentage}

\item{c}{connectivity}
}
\value{
list of Klemm matrices
}
\description{
Generates a list of Klemm-Eguiluz adjacency matrices
}
\details{
Generates a list compatible with other functions. For more details on Klemm-Eguiluz interaction matrices, please check the documentation of "generateA".
}
