% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSets.R
\name{generateSets}
\alias{generateSets}
\title{Generate multiple datasets}
\usage{
generateSets(n, klemms, species, samples, x, mode = "env", name)
}
\arguments{
\item{n}{number of replicates}

\item{x}{vector specificying environmental strength or removed species}

\item{mode}{"env" or "abundance", env takes environmental strength into account while "abundance" includes species removal}

\item{name}{filename of output dataset}
}
\value{
Returns nothing, but saves .rds files of datasets
}
\description{
Generate a list of list of datasets with n replicates for x datapoints.
}
\details{
Calls on the envGrowthChanges, generateDataSet and glv functions from seqtime to generate datasets compatible with other functions.
}
