% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_log_timeline_item.R
\name{log_timeline_item}
\alias{log_timeline_item}
\title{#' log_timeline_item}
\usage{
log_timeline_item(chglog_1)
}
\arguments{
\item{chglog_1}{A list of two item change and item}
}
\value{
timeline_item UI component
}
\description{
A convenient function to help render a single bs4Dash::timeline_item
}
\examples{
graph_chg_log = list(
  time = '2021-01-05',
  change = list(
    cmd = 'addEdge',
    from= 1,
    to = 2
  )
)
log_timeline_item(graph_chg_log)
}
