% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maximise.R
\name{maximize_helper}
\alias{maximize_helper}
\title{Javascript tha that help Resize Widget in `bs4Dash::box`}
\usage{
maximize_helper(id, offset = 180)
}
\arguments{
\item{id}{shiny id}

\item{offset}{number of pixel to offset from main screen}
}
\description{
Javascript tha that help Resize Widget in `bs4Dash::box`
}
