% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dynamic_attr_selector.R
\name{mod_dynamic_attr_selector_server}
\alias{mod_dynamic_attr_selector_server}
\title{dynamic_attr_selector Server Functions}
\usage{
mod_dynamic_attr_selector_server(
  id,
  dataInput,
  domain = getDefaultReactiveDomain()
)
}
\description{
dynamic_attr_selector Server Functions
}
