% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fileUploader.R
\name{mod_fileUploader_server}
\alias{mod_fileUploader_server}
\title{fileUploader Server Functions}
\usage{
mod_fileUploader_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
reactive igraph object
}
\description{
fileUploader Server Functions
}
