% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fileUploader.R
\name{mod_fileUploader_ui}
\alias{mod_fileUploader_ui}
\title{fileUploader UI Function}
\usage{
mod_fileUploader_ui(id, label = "Upload a File", ...)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{label}{message show before upload a file}

\item{...}{any other argument for `shiny::fileInput`}
}
\value{
bs4Dash UI
}
\description{
A shiny Module.
}
