% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visNet.R
\name{mod_visNet_server}
\alias{mod_visNet_server}
\title{visNet Server Functions}
\usage{
mod_visNet_server(id, graph_rct, debug = F, options, layout = layout)
}
\arguments{
\item{id}{shiny}

\item{graph_rct}{reactive expression of igraph}

\item{options}{list of option passed to `visSetOptions`}

\item{layout}{igraph layout to put in `visNetwork::visIgraphLayout`
mod_visNet_server links both modification and interaction UI and by doing so
make it easy}
}
\value{
list of `Current` and `Main`
`Main` for graph that has been committed editing;
`Current` for graph on display;
}
\description{
visNet Server Functions
}
