% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visInteraction.R
\name{mod_visNetInteraction_server}
\alias{mod_visNetInteraction_server}
\title{NetworkDisplayServer}
\usage{
mod_visNetInteraction_server(
  id,
  igraph_rct,
  e_ignore = c(),
  v_ignore = c(),
  show_hidden = F,
  domain = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{id}

\item{igraph_rct}{reactive expression for igraph}

\item{e_ignore}{a vector of edge attributes name to ignore}

\item{v_ignore}{a vector of node attributes name to ignore}

\item{domain}{session for when nesting module}
}
\description{
This module let you interact with graph
Require visnetwork rendered in shiny to have base id `visNetworkId`
}
