% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visNetworkRead.R
\name{mod_visNetworkReadControler_server}
\alias{mod_visNetworkReadControler_server}
\title{visNetworkReadControler Server Functions}
\usage{
mod_visNetworkReadControler_server(id, igraph_rct)
}
\arguments{
\item{id}{id}

\item{igraph_rct}{reactive expression for igraph}
}
\description{
visNetworkReadControler Server Functions
}
