% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visNetworkRead.R
\name{mod_visNetworkReadControler_ui}
\alias{mod_visNetworkReadControler_ui}
\title{visNetworkRead UI Function}
\usage{
mod_visNetworkReadControler_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
