% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modify_graph.R
\name{modify_graph_i}
\alias{modify_graph_i}
\title{modify_graph}
\usage{
modify_graph_i(
  igraphObj,
  visNetwork_graphChange,
  sideEffect = NULL,
  hard_delete = T
)
}
\arguments{
\item{igraphObj}{object that inherit igraph}

\item{visNetwork_graphChange}{a single list of command as a result of
visnetwork_graphChange, explained in their package
This list should always has following element:
- *addNode*
- *addEdge*
- *editEdge*
- *deleteElements*
If you ever need side effect, it is possible to add in your function
every time you make some change, add a new attribute to node are edge.
(for example a time stamp) so you can trick what has been changed and what
has not.}

\item{sideEffect}{this namespace is meant to be written down as a function
to trigger side effect different cmd type}

\item{hard_delete}{if set to false element will not be deleted but added an
attribute `.deleted_at` instead.}
}
\value{
a igraph object that has been changed
}
\description{
Modify graph from a single command line
}
