% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_networkUtilApp.R
\name{networkUtilApp}
\alias{networkUtilApp}
\title{networkUtilApp}
\usage{
networkUtilApp(graph)
}
\arguments{
\item{graph}{an igraph or tbl_graph object;
if you prefer node and edge notation call `igraph::graph_from_data_frame`
to convert node and edge sheet}
}
\value{
shiny session that quickly call you whatever graph
you put in into a graph
}
\description{
Explore your own graph and explort
}
