% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils-package.R
\docType{package}
\name{networkUtils-package}
\alias{networkUtils}
\alias{networkUtils-package}
\title{networkUtils: networkUtils}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: Fanzhou Liang \email{fzlianguk@gmail.com}

}
\keyword{internal}
