% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pasteDetails.R
\name{pasteDetails}
\alias{pasteDetails}
\alias{pasteNodeDetails}
\alias{pasteEdgeDetails}
\title{pasteDetails}
\usage{
pasteDetails(
  namedList,
  include,
  sep.attr = "<br>",
  sep.label = ":",
  show_hidden = F
)

pasteNodeDetails(g, include, ...)

pasteEdgeDetails(g, include, ...)
}
\arguments{
\item{namedList}{a named r list}

\item{include}{a vector of names from much you want to include}

\item{sep.attr}{this is typically line breaks in between each attribute}

\item{sep.label}{the speperator between attribute name and attributes}

\item{show_hidden}{attributes begain with '.' will be pasted into this.}

\item{g}{igraph object specifically used for graph objects}
}
\value{
The return value, if any, from executing the function.
}
\description{
paste list of vector (same length) together.
You can use these function to write html paragraph for simple just one row
Use-full with graph node and edge attributes.
}
\details{
Paste attributes into a rectangular tools

exclusivedly works for pasting
}
\examples{
# example code

# paste normal object
nL = list(
 id = 1,
  object = "Apple",
  category = "Fruit",
  size = 24.22
)
x_html = pasteDetails(nL)
cat(x_paste)
}
