% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHubs.R
\name{plotHubs}
\alias{plotHubs}
\title{Generate SuperExactTest-compatible dataset}
\usage{
plotHubs(hubs, toolnames)
}
\arguments{
\item{toolnames}{list of toolnames}

\item{list.hubs}{list of hubs per tool}

\item{mode}{"hub" or "central": for hub, the degree is returned, for central, the centrality scores}
}
\value{
Dataframe with hub or central species for tools
}
\description{
Takes list of hub species or highest centrality species and converts into SuperExactTest-compatible vectors
}
\details{
Includes all possible sets.
}
