% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrecisionvsRecall.R
\name{plotPrecisionvsRecall}
\alias{plotPrecisionvsRecall}
\title{Plot precision vs recall}
\usage{
plotPrecisionvsRecall(results, toolnames, x, colours, mode = "env")
}
\arguments{
\item{results}{output statistics per tool}

\item{toolnames}{list of toolnames}

\item{x}{vector specificying environmental strength or removed species}

\item{colours}{vector specificying codes in hex}
}
\value{
ggplot object
}
\description{
Produces ggplot with precision vs sensitivity scatterplot, coloured by tool.
}
\details{
Requires the output statistics file, not a network file.
}
