% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeries.R
\name{plotSeries}
\alias{plotSeries}
\title{Plot statistics}
\usage{
plotSeries(results, klemmspecs, toolnames, x, statistic, axis,
  mode = "linear", legend = TRUE, colours)
}
\arguments{
\item{results}{output statistics per tool}

\item{klemmspecs}{output statistics for Klemm-Eguiluz matrices}

\item{toolnames}{list of toolnames}

\item{x}{vector specificying environmental strength or removed species}

\item{statistic}{statistic to be plotted}

\item{axis}{x axis title}

\item{mode}{"linear" or "quadratic" for trend line}

\item{legend}{if TRUE, show legend}

\item{colours}{vector specificying codes in hex}
}
\value{
ggplot object
}
\description{
Produces ggplot with specified statistic, coloured by tool.
}
\details{
Requires the output statistics file, not a network file. Statistics that can be plotted are:
"Precision", "Sensitivity", "Specificity", "AUC", "Transitivity", "Degree", "Scale-freeness"
}
