% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Generate SuperExactTest p-value vector corrected for sampling design}
\usage{
plotStats(allvecs)
}
\arguments{
\item{allvecs}{list of hubs per matrix / tool, plotHubs output}
}
\value{
Vector with p-values that can be used to modify SuperExactTest object
}
\description{
Takes list of hub species or highest centrality species and converts into p-values
}
\details{
Includes all possible sets.
}
