% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSuperSet.R
\name{plotSuperSet}
\alias{plotSuperSet}
\title{Generate ggplots using some SuperExactTest code}
\usage{
plotSuperSet(hubs, toolnames)
}
\arguments{
\item{toolnames}{list of toolnames}

\item{list.hubs}{list of hubs per tool}

\item{mode}{"hub" or "central": for hub, the degree is returned, for central, the centrality scores}
}
\value{
Dataframe with hub or central species for tools
}
\description{
Takes list of hub species or highest centrality species and outputs heatmap
}
\details{
Includes all possible sets.
}
