% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUpset.R
\name{plotUpset}
\alias{plotUpset}
\title{Generate UpSetplot-compatible dataset}
\usage{
plotUpset(list.hubs, list.toolnames)
}
\arguments{
\item{list.hubs}{list of hubs per tool}

\item{toolnames}{list of toolnames}
}
\value{
Dataframe with hub or central species for tools
}
\description{
Takes list of hub species or highest centrality species and converts into upset-compatible plot.
}
