% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoNet.R
\name{readCoNet}
\alias{readCoNet}
\title{Read CoNet outputs}
\usage{
readCoNet(name = "brown", mode = "env", x = seq(from = 0, to = 50, by =
  5), n = 10, alldata = alldata, setname, tool = tool, klemmadj, wdir)
}
\arguments{
\item{name}{name of p-value merge, "brown" or "fisher"}

\item{x}{vector specificying environmental strength or removed species}

\item{n}{number of replicates}

\item{alldata}{list of datasets with simulated counts}

\item{setname}{name of simulation}

\item{tool}{toolname, pref "CoNetBrown" or "CoNetFisher"}

\item{klemmadj}{list of Klemm matrices}
}
\value{
nothing, but writes .txt files to current directory
}
\description{
Reads CoNet .txt outputs and converts to adjacency matrix.
}
\details{
CoNet bash calls output a text file that can be used to compute adjacency matrices. This function parses the text file.
}
