% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpar.R
\name{readSpar}
\alias{readSpar}
\title{Read SparCC outputs}
\usage{
readSpar(n, x, mode = NA, setname, alldata, klemmadj, wdir)
}
\arguments{
\item{n}{number of replicates}

\item{x}{vector specificying environmental strength or removed species}

\item{setname}{name of simulation}

\item{alldata}{list of datasets with simulated counts}

\item{klemmadj}{list of Klemm matrices}
}
\value{
nothing, but writes .txt files to current directory
}
\description{
Reads SparCC .txt outputs and converts to adjacency matrix.
}
\details{
SparCC bash calls output p-values that can be used to compute adjacency matrices.
}
