% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_search_idx.R
\name{search_idx}
\alias{search_idx}
\title{Search index}
\usage{
search_idx(g, expr_txt, as_ids = F, search_in = "nodes")
}
\arguments{
\item{g}{igraph object}

\item{expr_txt}{string of expression or names of node to search}

\item{as_ids}{when T return only index}

\item{search_in}{dip argument that when set to edge can search in edges}
}
\value{
a vector
}
\description{
A utils function
}
\examples{
set.seed(1)
g = make_tree(12, 3)
V(g)$attr <- sample(letters, length(g))
V(g)$name <- paste0(seq(length(g)), ".", V(g)$attr)
search_idx(g, "attr == 'v'")
search_idx(g, '9.w')
# if tidygraph is loaded in environment this will work
library(tidygraph)
search_idx(g, 'node_is_root()')

}
