% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinySpider-package.R
\docType{package}
\name{ShinySpider-package}
\alias{ShinySpider}
\alias{ShinySpider-package}
\title{ShinySpider: PKG_TITLE}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
