% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleNetworkUtilApp_server.R
\name{simpleNetworkUtilApp_server}
\alias{simpleNetworkUtilApp_server}
\title{The application server-side}
\usage{
simpleNetworkUtilApp_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.
DO NOT REMOVE.}
}
\description{
The application server-side
}
