% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testgCoda.R
\name{testgCoda}
\alias{testgCoda}
\title{Calls gCoda function}
\usage{
testgCoda(data)
}
\arguments{
\item{data}{input dataset}
}
\value{
adjacency matrix inferred from data
}
\description{
Calls gCoda and formats output to adjacency matrix
}
\details{
Author: Fang Huaying. gCoda is available at: https://github.com/huayingfang/gCoda
}
