% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSpear.R
\name{testSpear}
\alias{testSpear}
\title{Calls Spearman correlation}
\usage{
testSpear(data)
}
\arguments{
\item{data}{input dataset}
}
\value{
adjacency matrix inferred from data
}
\description{
Calls Spearman correlation as implemented in CoNet and formats output to adjacency matrix
}
\details{
Author: Karoline Faust.
}
