% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSpiec.R
\name{testSpiec}
\alias{testSpiec}
\title{Calls SpiecEasi}
\usage{
testSpiec(data, method = "mb")
}
\arguments{
\item{data}{input dataset}

\item{method}{"mb" or "glasso", please check SpiecEasi's documentation for more information}
}
\value{
adjacency matrix inferred from data
}
\description{
Calls SpiecEasi and formats output to adjacency matrix
}
\details{
Authors: Zachary D. Kurtz et al. SpiecEasi is available at: https://github.com/zdk123/SpiecEasi
}
