% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vis_fibre.R
\name{vis_fibre}
\alias{vis_fibre}
\title{vis_fibre}
\usage{
vis_fibre(tg, treeSpacing = 90, nodeSpacing = 50, select_by = "enclosure")
}
\arguments{
\item{tg}{a tree graph generated by auto splice}

\item{treeSpacing, nodeSpacing}{inherited graph parameter from visIgraph}

\item{select_by}{default enclosure, recommend fibre_spn
This will select highlight only fibre of a certain spine node area.}
}
\value{
visNetwork graph object
}
\description{
A fct function
This function visualize a graph that is generated by function `split_tree()`
and `index_fibre()`. To visualise FAS sheet, special data transfromation will
be needed.

By default automatically generated FAS will re-index each node based on leaf index.
For example "FDP/1234-1" indicated this fibre node is going to the first splitter
The appendix "-1" does not have much meanings. But to avoid misleading readers,
`vis_fibre()` will automatically correct name of each fibre node. For cabinet,
the new appendix inherit the *fibre_to* fibre index. For other nodes, they inherits
*fibre_from* index (as this will make sure each node is unique).
}
