% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visModification.R
\name{visNetModification}
\alias{visNetModification}
\alias{mod_visNetModification_ui}
\alias{mod_visNetModification_server}
\title{visNetworkWrite UI Function}
\usage{
mod_visNetModification_ui(id)

mod_visNetModification_server(
  id,
  igraphObj,
  dev = F,
  hard_delete = T,
  NodeAttrTooltip = T,
  EdgeAttrTooltip = T,
  domain = getDefaultReactiveDomain(),
  visNet_options = NULL,
  layout = "layout_nicely"
)
}
\arguments{
\item{id}{shiny server id}

\item{igraphObj}{a reactive graph object}

\item{dev}{open a printer}

\item{hard_delete}{when this flag is turn off, no edge can be actually deleted/}

\item{NodeAttrTooltip, EdgeAttrTooltip}{when these two flag are set to ture
attributes will be automatically parsed into tooltips, following formula
'attrName: attr'. There is no good parsing for time series type.}

\item{domain}{session}

\item{visNet_options}{list of option passed to `visSetOptions`}

\item{layout}{igraph layout to put in `visNetwork::visIgraphLayout`}
}
\value{
reactiveValues $Curent and $Main and more
}
\description{
A shiny Module.
}
\details{
$Current is a reactive igraph Object that every is being modified now
$Main is the igraph Object that has been committed and saved
In addition it return a set of `reactiveValues` which monitor graph changes
and track node that is currently clicked.
}
