% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFeatures.R
\name{writeFeatures}
\alias{writeFeatures}
\title{Write features}
\usage{
writeFeatures(n, x, edges)
}
\arguments{
\item{n}{number of replicates}

\item{x}{vector specificying environmental strength or removed species}

\item{dataset}{full dataset}
}
\value{
nothing, but writes .txt files to current directory
}
\description{
Writes features to txt files.
}
\details{
List of list of features can be written to text files for use by CoNet bash calls.
}
