% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSets.R
\name{writeSets}
\alias{writeSets}
\title{Write datasets}
\usage{
writeSets(n, x, dataset)
}
\arguments{
\item{n}{number of replicates}

\item{x}{vector specificying environmental strength or removed species}

\item{dataset}{full dataset}
}
\value{
nothing, but writes .txt files to current directory
}
\description{
Writes data sets to txt files.
}
\details{
List of list of datasets can be written to text files for use by SparCC or CoNet bash calls.
}
