% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visInteraction.R
\name{mod_visNetInteraction_server}
\alias{mod_visNetInteraction_server}
\title{NetworkDisplayServer}
\usage{
mod_visNetInteraction_server(
  id,
  igraph_rct,
  e_ignore = c(),
  v_ignore = c(),
  show_hidden = F,
  domain = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{id}

\item{igraph_rct}{reactive expression for igraph}

\item{e_ignore, v_ignore}{a vector of node/edge attributes. This tells the
interactive UI to stop parsing specific attribute into searchable UI.}

\item{show_hidden}{used in conjunction with `e_ignore` and `v_ignore`. If you
have an attribute begain/has with '.' This flag will show it when set to TRUE.
This is useful against attributes created by tidygraph (`.tidygraph_e_index`)
or when you have to create color attributes for visNetwork (`color.border`)}

\item{domain}{session for when nesting module}
}
\description{
This module let you interact with graph
Require visnetwork rendered in shiny to have base id `visNetworkId`
}
