% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visNet.R
\name{mod_visNet_}
\alias{mod_visNet_}
\alias{mod_visNet_server}
\title{visNet Server Functions}
\usage{
mod_visNet_server(
  id,
  graph_rct,
  debug = F,
  visNet_options = NULL,
  layout = NULL,
  NodeAttrTooltip = F,
  EdgeAttrTooltip = F
)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{graph_rct}{reactive expression of igraph}

\item{debug}{this flag will create a text output for tracking change.}

\item{visNet_options}{list of option passed to `visSetOptions`
this could also be reactive}

\item{layout}{igraph layout to put in `visNetwork::visIgraphLayout`
could be static or reactive.}

\item{NodeAttrTooltip, EdgeAttrTooltip}{when these two flag are set to ture
attributes will be automatically parsed into tooltips, following formula
'attrName: attr'. There is no good parsing for time series type.}
}
\value{
reactive list
}
\description{
The easiest way to use both modification server and interaction server.
mod_visNet_server links both modification and interaction UI and by doing so
make it easy
}
\details{
list of `Current`, `Main` `click_node` and `click_edge`
`Main` for graph that has been committed editing;
`Current` for graph on display;
Two additional `click_node` and `click_edge` let you track
edges and node currently clicked
}
