% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visInteraction.R
\name{mod_visNetInteraction_server}
\alias{mod_visNetInteraction_server}
\title{visNetworkReadControler Server Functions}
\usage{
mod_visNetInteraction_server(id, igraph_rct)
}
\arguments{
\item{id}{id}

\item{igraph_rct}{reactive expression for igraph}
}
\description{
This module let you interact with graph
Require visnetwork rendered in shiny to have base id `visNetworkId`
}
