% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visModification.R
\name{mod_visNetModification_server}
\alias{mod_visNetModification_server}
\title{visNetworkWrite Server Functions}
\usage{
mod_visNetModification_server(
  id,
  igraphObj,
  dev = F,
  hard_delete = T,
  NodeAttrTooltip = T,
  EdgeAttrTooltip = T,
  domain = getDefaultReactiveDomain(),
  options = NULL
)
}
\arguments{
\item{id}{shiny server id}

\item{igraphObj}{a reactive graph object}

\item{domain}{session for}

\item{options}{list of option passed to `visSetOptions`}
}
\value{
reactiveValues $Curent and $Main
}
\description{
visNetworkWrite Server Functions
}
\details{
$Current is a reactive igraph Object that every is being modified now
$Main is the igraph Object that has been committed and saved
}
