% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modify_graph.R
\name{modify_graph_i}
\alias{modify_graph_i}
\title{modify_graph}
\usage{
modify_graph_i(igraphObj, visNetwork_graphChange, sideEffect = NULL)
}
\arguments{
\item{igraphObj}{object that inherit igraph}

\item{sideEffect}{this namespace is meant to be written down as a function
to trigger side effect different cmd type}

\item{visNetwork_graphChage}{a single list of command as a result of
visnetwork_graphChange, explained in their package}
}
\value{
The return value, if any, from executing the function.
}
\description{
Modify graph from a single command line
}
