% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkUtils-package.R
\docType{package}
\name{networkUtils-package}
\alias{networkUtils}
\alias{networkUtils-package}
\title{networkUtils: Interactive Extension for VisNetwork}
\description{
Shiny extension for Almende B.V. 's package `visNetwork.`networkUtils let you modify networkgraph or downloadnode and edge sheets. The returned igraph object can be used for any purpose.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FzliangFrank/networkUtils}
  \item \url{https://fzliangfrank.github.io/networkUtils/}
  \item Report bugs at \url{https://github.com/FzliangFrank/networkUtils/issues}
}

}
\author{
\strong{Maintainer}: Fanzhou Liang \email{fzlianguk@gmail.com}

}
\keyword{internal}
